+++ src/ipaperftest/core/__init__.py
@@ -1,0 +1,5 @@
+#
+# Copyright (C) 2021 FreeIPA Contributors see COPYING for license
+#
+
+__import__('pkg_resources').declare_namespace(__name__)
